ml/requirements-docker.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Machine learning
scikit-learn>=1.3.0
imbalanced-learn>=0.12.0
statsmodels>=0.14.0
xgboost>=2.0.0
lightgbm>=4.3.0
hmmlearn>=0.3.0
arch>=6.0.0
# Note: tensorflow-cpu removed for Docker compatibility on ARM64
# TensorFlow is disabled in production via ENABLE_TRANSFORMER=False

# Data sources
supabase>=2.0.0
yfinance>=0.2.0
alpaca-py>=0.8.0
databento>=0.40.0
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Database
psycopg2-binary>=2.9.0

# Caching
redis>=4.5.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Visualization (optional but recommended)
plotly>=5.18.0

# FastAPI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
